app/requirements.txt
Flask==2.0.2
cryptography==35.0.0
SQLAlchemy==1.4.27
email-validator==1.1.3
Flask-RESTful==0.3.9
redis==4.0.1
PyMySQL==1.0.2
marshmallow==3.14.1
PyJWT==2.3.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-HTTPAuth==4.5.0
